tablevault/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "tablevault"
version = "0.2.0"
description = "TableVault - A centralized data repository for cross-process data filtering"
requires-python = ">=3.8"
dependencies = [
    "python-arango",
]

[project.optional-dependencies]
dev = [
    "pytest",
]

[tool.setuptools.packages.find]
include = ["tablevault*"]
